+++ src/main/java/si/matjazcerkvenik/alertmonitor/web/UiReportBean.java
@@ -1,33 +1,49 @@
 package si.matjazcerkvenik.alertmonitor.web;
 
+import org.primefaces.model.charts.ChartData;
+import org.primefaces.model.charts.line.LineChartDataSet;
+import org.primefaces.model.charts.line.LineChartModel;
+import org.primefaces.model.charts.line.LineChartOptions;
+import org.primefaces.model.charts.optionconfig.title.Title;
 import si.matjazcerkvenik.alertmonitor.model.prometheus.PQueryMessage;
-import si.matjazcerkvenik.alertmonitor.model.prometheus.PrometheusApi;
+import si.matjazcerkvenik.alertmonitor.model.prometheus.PrometheusApiClient;
+import si.matjazcerkvenik.alertmonitor.model.prometheus.PrometheusApiClientPool;
+import si.matjazcerkvenik.alertmonitor.util.Formatter;
 import si.matjazcerkvenik.alertmonitor.util.LogFactory;
 
 import javax.faces.bean.ManagedBean;
 import javax.faces.bean.SessionScoped;
+import java.text.DecimalFormat;
+import java.util.ArrayList;
+import java.util.List;
 
 @ManagedBean
 @SessionScoped
 public class UiReportBean {
 
+    private final String QUERY_PROM_UP_TIME = "round(time()-process_start_time_seconds{job=\"prometheus\"})";
     private final String QUERY_COUNT_TARGETS_ALL = "count(up)";
     private final String QUERY_COUNT_TARGETS_DOWN = "count(up==0) + count(probe_success==0)";
-    private final String QUERY_AVERAGE_TARGETS_AVAILABILITY = "avg(avg_over_time(up[1h])) * 100";
+    private final String QUERY_AVERAGE_TARGETS_AVAILABILITY = "(sum(avg_over_time(up[1h])) + sum(avg_over_time(probe_success[1h])) - count(probe_success)) / (count(up) - count(probe_success)) * 100";
+    private final String QUERY_TARGETS_LIVENESS = "(count(up == 1) + count(probe_success == 1) - count(probe_success)) / (count(up) - count(probe_success)) * 100";
     private final String QUERY_COUNT_JOBS_ALL = "count(count(up) by (job))";
+    private final String QUERY_PROMETHEUS_AVERAGE_REQUEST_DURATION_MILLIS_TEMPLATE = "sum(rate(prometheus_http_request_duration_seconds_sum[__INTERVAL__m])) / sum(rate(prometheus_http_request_duration_seconds_count[__INTERVAL__m])) * 1000";
+    private final String QUERY_PROMETHEUS_90_PERCENT_REQUEST_DURATION = "histogram_quantile(0.90, sum(rate(prometheus_http_request_duration_seconds_bucket[10m])) by (le)) * 1000";
 
-    private final String QUERY_PROMETHEUS_AVERAGE_REQUEST_DURATION_MILLIS_TEMPLATE = "sum(rate(prometheus_http_request_duration_seconds_sum[{__INTERVAL__}m])) / sum(rate(prometheus_http_request_duration_seconds_count[{__INTERVAL__}m])) * 1000";
-    private final String QUERY_PROMETHEUS_AVERAGE_REQUEST_DURATION_MILLIS_1 = "sum(rate(prometheus_http_request_duration_seconds_sum[1m])) / sum(rate(prometheus_http_request_duration_seconds_count[1m])) * 1000";
-    private final String QUERY_PROMETHEUS_AVERAGE_REQUEST_DURATION_MILLIS_5 = "sum(rate(prometheus_http_request_duration_seconds_sum[5m])) / sum(rate(prometheus_http_request_duration_seconds_count[5m])) * 1000";
-    private final String QUERY_PROMETHEUS_AVERAGE_REQUEST_DURATION_MILLIS_15 = "sum(rate(prometheus_http_request_duration_seconds_sum[15m])) / sum(rate(prometheus_http_request_duration_seconds_count[15m])) * 1000";
+
+    public String getPrometheusUpTime() {
+        PQueryMessage queryMessage = executeQuery(QUERY_PROM_UP_TIME);
+        if (queryMessage != null) {
+            int t = Integer.parseInt(queryMessage.getData().getResult().get(0).getValue()[1].toString());
+            return Formatter.convertToDHMSFormat(t);
+        }
+        return "n/a";
+    }
 
     public String getCountTargetsDown() {
-        try {
-            PrometheusApi api = new PrometheusApi();
-            PQueryMessage queryMessage = api.query(QUERY_COUNT_TARGETS_DOWN);
+        PQueryMessage queryMessage = executeQuery(QUERY_COUNT_TARGETS_DOWN);
+        if (queryMessage != null) {
             return queryMessage.getData().getResult().get(0).getValue()[1].toString();
-        } catch (Exception e) {
-            LogFactory.getLogger().error("UiReportBean: getCountTargetsDown: exception: ", e);
         }
         return "n/a";
     }
@@ -35,7 +51,17 @@ public class UiReportBean {
     public String getAverageTargetAvailability() {
         PQueryMessage queryMessage = executeQuery(QUERY_AVERAGE_TARGETS_AVAILABILITY);
         if (queryMessage != null) {
-            return queryMessage.getData().getResult().get(0).getValue()[1].toString();
+            Double d = Double.parseDouble(queryMessage.getData().getResult().get(0).getValue()[1].toString());
+            return new DecimalFormat("0.00").format(d);
+        }
+        return "n/a";
+    }
+
+    public String getTargetsLiveness() {
+        PQueryMessage queryMessage = executeQuery(QUERY_TARGETS_LIVENESS);
+        if (queryMessage != null) {
+            Double d = Double.parseDouble(queryMessage.getData().getResult().get(0).getValue()[1].toString());
+            return new DecimalFormat("0.00").format(d);
         }
         return "n/a";
     }
@@ -57,19 +83,105 @@ public class UiReportBean {
     }
 
     public String getPrometheusAverageRequestDuration(int historyMinutes) {
-        PQueryMessage queryMessage = executeQuery(QUERY_PROMETHEUS_AVERAGE_REQUEST_DURATION_MILLIS_TEMPLATE.replace("{__INTERVAL__}", ""+historyMinutes));
+        PQueryMessage queryMessage = executeQuery(QUERY_PROMETHEUS_AVERAGE_REQUEST_DURATION_MILLIS_TEMPLATE.replace("__INTERVAL__", ""+historyMinutes));
         if (queryMessage != null) {
-            return queryMessage.getData().getResult().get(0).getValue()[1].toString();
+            Double d = Double.parseDouble(queryMessage.getData().getResult().get(0).getValue()[1].toString());
+            return new DecimalFormat("0.00").format(d);
         }
         return "n/a";
     }
 
+    public String getPrometheus90PercentRequestDuration() {
+        PQueryMessage queryMessage = executeQuery(QUERY_PROMETHEUS_90_PERCENT_REQUEST_DURATION);
+        if (queryMessage != null) {
+            Double d = Double.parseDouble(queryMessage.getData().getResult().get(0).getValue()[1].toString());
+            return new DecimalFormat("0.00").format(d);
+        }
+        return "n/a";
+    }
+
+
+    private LineChartModel availabilityLineModel;
+
+    private LineChartModel createSingleValueLineModel(String chartTitle, LineChartModel model, PQueryMessage queryMessage) {
+        model = new LineChartModel();
+        ChartData data = new ChartData();
+
+        List<Object[]> result = queryMessage.getData().getResult().get(0).getValues();
+
+        LineChartDataSet dataSet = new LineChartDataSet();
+        List<Object> values = new ArrayList<>();
+        for (int i = 0; i < result.size(); i++) {
+            values.add(Double.parseDouble(result.get(i)[1].toString()));
+        }
+        dataSet.setData(values);
+        dataSet.setFill(false);
+        dataSet.setLabel("Dataset");
+        dataSet.setBorderColor("rgb(75, 192, 192)");
+        dataSet.setTension(0.01);
+        data.addChartDataSet(dataSet);
+
+        List<String> labels = new ArrayList<>();
+        for (int i = 0; i < result.size(); i++) {
+            labels.add(result.get(i)[0].toString());
+        }
+        data.setLabels(labels);
+
+        //Options
+        LineChartOptions options = new LineChartOptions();
+        Title title = new Title();
+        title.setDisplay(true);
+        title.setText(chartTitle);
+        options.setTitle(title);
+
+        model.setOptions(options);
+        model.setData(data);
+
+        return model;
+    }
+
+    public LineChartModel getAvailabilityLineModel() {
+        PQueryMessage queryMessage = executeQueryRange(QUERY_AVERAGE_TARGETS_AVAILABILITY);
+        if (queryMessage != null) {
+            availabilityLineModel = createSingleValueLineModel("Availability", availabilityLineModel, queryMessage);
+        }
+        return availabilityLineModel;
+    }
+
+
+    private LineChartModel livenessLineModel;
+
+    public LineChartModel getLivenessLineModel() {
+        PQueryMessage queryMessage = executeQueryRange(QUERY_TARGETS_LIVENESS);
+        if (queryMessage != null) {
+            livenessLineModel = createSingleValueLineModel("Liveness", livenessLineModel, queryMessage);
+        }
+        return livenessLineModel;
+    }
+
+
+
     private PQueryMessage executeQuery(String query) {
+        PrometheusApiClient api = PrometheusApiClientPool.getInstance().getClient();
         try {
-            PrometheusApi api = new PrometheusApi();
             return api.query(query);
         } catch (Exception e) {
             LogFactory.getLogger().error("UiReportBean: executeQuery: " + query + ": exception: ", e);
+        } finally {
+            PrometheusApiClientPool.getInstance().returnClient(api);
+        }
+        return null;
+    }
+
+    private PQueryMessage executeQueryRange(String query) {
+        PrometheusApiClient api = PrometheusApiClientPool.getInstance().getClient();
+        try {
+            long t = System.currentTimeMillis() / 1000;
+            return api.queryRange(query, (t - 3600), t, "1m");
+        } catch (Exception e) {
+            LogFactory.getLogger().error("UiReportBean: executeQueryRange: " + query + ": exception: ", e);
+        } finally {
+            PrometheusApiClientPool.getInstance().returnClient(api);
         }
         return null;
     }